Manufactura Independente's edits for a design update
* Streamlined look on the front page and MP page (unfinished)
* New fonts for the site (Playfair Display and Nobile)
* New logo
* Party colours now in CSS, instead of coded in PHP. We might have missed
some PHP templates in this transition. This change was motivated by our
use of Sass for developing
* Included Foundation 6
* Updated jQuery to v.2.2.4 (compatible with Foundation) for the jquery menus
and other interface perks of the Foundation framework
* Included fondation-icons as a quick alternative for updating icons in the
website

We worked on a local copy using gulp to compile sass files and included the
minified result at app.css. This was a quick and dirty solution for the
hackathon work. We can provide the scss files and it would be great to continue
the work on css using the sass pipeline.